pyproject.toml
[project]
name = "claude-apps"
version = "1.0.0"
description = "Consolidated Python applications for Claude Code"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34.0",
    "gitpython>=3.1.0",
    "InquirerPy>=0.3.0",
    "loguru>=0.7.0",
    "playwright>=1.40.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "ruamel.yaml>=0.18.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "moto>=5.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps/src/claude_apps"]

[tool.hatch.build.targets.sdist]
include = ["apps/src/claude_apps"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["apps/src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["apps/tests"]
pythonpath = ["apps/src"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["apps/src/claude_apps"]
omit = ["apps/tests/*"]
